+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## next
 
 - Fixed missed error on progressbar's value when the progressbar finishes with error
+- Fixed value counting in the `signature` view when a property contains duplicated objects, which previously led to incorrect `optional` label
 - Fixed warning "Added non-passive event listener..." in Chrome caused by CodeMirror (https://github.com/codemirror/codemirror5/issues/6735)
 - Removed missed `console.log()` in progressbar