+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 1.0.0-beta.87 (15-10-2024)
 
 - Enhanced style of errors on the loading data overlay
 - Enhanced responsiveness and addressing issues with freezing during transition states for views related to the data loading state. Instead of setup of initial state via JavaScript, the new `@starting-style` at-rule is used: